reorganize test directory to use multiple suites
