+++ __tests__/testData/skyrim-test-entries.json
